Create df_parser.py
Data parser for excel that is built for large spreadsheets with several categories that will seek out a series of categories.  For example, maybe you have a spreadsheet of music albums by years and you want to pull all the albums from 1995, this tool will pull all the albums of 1995 into a data frame so it's easier to sort/locate information.

The second part of the code will loop through and print the categories to help you search. Maybe you forgot what years you have stored, use this tool to print them out without dupes.